+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## [Unreleased]
 
 ### Changed
-- As per this PR - https://github.com/vtex-apps/shopper-location/pull/51, Shopper Location now will show province has dropdown field. So, updated the testcase
+- Handle both province dropdown and text field
+
+### Changed
+- Added a link for Notion page on errors
 
 ### Changed
 - Use Sync Checkout UI Custom via API everywhere